Force feature unification between payjoin-test-utils and payjoin
