+++ payjoin-test-utils/Cargo.toml
@@ -8,22 +8,40 @@ repository = "https://github.com/payjoin/rust-payjoin"
 rust-version = "1.85"
 license = "MIT"
 
+[features]
+default = ["v2"]
+v2 = [
+  "payjoin/io",
+  "payjoin/_manual-tls",
+  "dep:axum-server",
+  "dep:ohttp",
+  "dep:rcgen",
+  "dep:rustls",
+  "dep:reqwest",
+  "dep:http",
+  "dep:tempfile",
+]
+
 [dependencies]
-axum-server = { version = "0.8", features = ["tls-rustls-no-provider"] }
+axum-server = { version = "0.8", features = [
+  "tls-rustls-no-provider",
+], optional = true }
 bitcoin = { version = "0.32.7", features = ["base64"] }
 corepc-node = { version = "0.10.0", features = ["download", "29_0"] }
-http = "1.3.1"
-ohttp = { package = "bitcoin-ohttp", version = "0.6.0" }
+http = { version = "1.3.1", optional = true }
+ohttp = { package = "bitcoin-ohttp", version = "0.6.0", optional = true }
 once_cell = "1.21.3"
-payjoin = { version = "0.25.0", features = ["io", "_manual-tls"] }
+payjoin = { version = "0.25.0", default-features = false }
 payjoin-mailroom = { path = "../payjoin-mailroom", features = ["_manual-tls"] }
 # Pin to 0.14.7 (last version to support our MSRV 1.85)
-rcgen = "=0.14.7"
+rcgen = { version = "=0.14.7", optional = true }
 reqwest = { version = "0.12.23", default-features = false, features = [
   "rustls-tls",
-] }
-rustls = { version = "0.23.31", default-features = false, features = ["ring"] }
-tempfile = "3.20.0"
+], optional = true }
+rustls = { version = "0.23.31", default-features = false, features = [
+  "ring",
+], optional = true }
+tempfile = { version = "3.20.0", optional = true }
 # Pin to 0.3.45 (last version to support our MSRV 1.85)
 time = "=0.3.45"
 tokio = { version = "1.47.1", features = ["full"] }
