+++ payjoin/Cargo.toml
@@ -33,7 +33,7 @@ _core = [
 ]
 directory = []
 v1 = ["_core"]
-v2 = ["_core", "hpke", "bhttp", "ohttp", "directory"]
+v2 = ["_core", "hpke", "bhttp", "ohttp", "directory", "payjoin-test-utils/v2"]
 #[doc = "Functions to fetch OHTTP keys via CONNECT proxy using reqwest. Enables `v2` since only `v2` uses OHTTP."]
 io = ["v2", "reqwest/rustls-tls"]
 _manual-tls = ["rustls"]
@@ -63,7 +63,7 @@ ignored = ["bitcoin-units"]
 
 [dev-dependencies]
 once_cell = "1.21.3"
-payjoin-test-utils = { version = "0.0.1" }
+payjoin-test-utils = { version = "0.0.1", default-features = false }
 tokio = { version = "1.47.1", features = ["full"] }
 tracing = "0.1.41"
